First working version of ka-clone tool
Summary:
I'll intentionally omit any description here because as a new tool,
I want to make sure the README and documentation are sufficient to
explain purpose.

The only color I will add for internal review is this will probably
first be used for khan-dotfiles repo cloning, but can also be a good
general purpose userspace tool for those of us who do a lot of things
manually.

Note the plaintext-ish examples in the examples directory also
function as integration tests(!) via `make tests`, requires Cram to be
installed which is available in dev-requirements.txt

Test Plan: Covered via integration tests

Reviewers: ethan, marcos, alpert, csilvers

Reviewed By: csilvers

Subscribers: ben, kevindangoor

Differential Revision: https://phabricator.khanacademy.org/D18395