fix: handle nonexistent git hooks directory
Summary:
It is possible for a .git subdirectory of a repository to not contain
the default hooks/ directory (perhaps the user deleted it, for example).

In this situation, we should just gracefully recreate it, rather than
dying.

Test Plan: delete .git/hooks in test repository, verify ka-clone —repair fixes it.

Reviewers: ethan, csilvers

Reviewed By: csilvers

Subscribers: anastassia

Differential Revision: https://phabricator.khanacademy.org/D18826