Add a hook to prevent replaying (changing the sha1) of commits in master.
Summary:
The way 'git rebase <foo>' works is it takes all commits that are on
your branch but not on foo, and 'replays' them on top of 'foo'.
This causes those commits to have a new sha1.

While doing this should never cause any problems, it does cause us
annoyance because when the replayed commits include a phabricator
audit, it causes a new audit to be created.  It also makes it more
difficult to talk about a commit, since the same commit might have
two different sha1's.

For these reasons, we disallow replaying commits that are on master.
Since we only commit stuff to master that's been deployed, this
means you can not replay any already-live code.

Use the simpler --abbrev-ref.

Oops, needed to fix permissions.

Print the commits that are in master that would be replayed.

I had to switch to bash to do this, so I could use process
substitution with comm.

Simplify the error-condition test now.

Test Plan:
I created a test repo (git clone git@github.com:csilvers/uap-core)
and ran the following set of commands:

   git co master
   git branch -D mytest
   git branch -D mytest2
   git co -b mytest
   git reset --hard HEAD^
   git co -b mytest2
   echo >> README.md
   git commit -a -m wip2
   git co mytest
   echo >> LICENSE
   git commit -a -m wip1
   git merge master
   echo >> LICENSE
   git commit -a -m wip3
   git rebase mytest2

It said:
```
Disabling this rebase: it would replay commits that are already in master.
Try changing to branch 'mytest' and doing a git push,
and then running this command again.
```

(same)

(none new)

Same, this time the error message was:

  Disabling this rebase: it would replay commits that are already in master:
     146cd3b
  Try changing to branch 'mytest' and doing a git push,
  and then running this command again.

and when I changed the '-eq' to 1 (from 8), it said:

  Disabling this rebase: it would replay commits that are already in master:
     146cd3b ...
  Try changing to branch 'mytest' and doing a git push,
  and then running this command again.

(same)

Reviewers: mroth, neel, jlfwong

Reviewed By: jlfwong

Subscribers: davidwang

Differential Revision: https://phabricator.khanacademy.org/D20231