fix install git hooks for git submodules
Summary:
git submodules use a GITDIR that is actually nested in the GITDIR of
their parent repository.  This fix looks up the proper GITDIR before
attempting to copy git hooks so that this process will work when in a
git submodule.

We still need to investigate a proper holistic solution for cloning our
mega-repo that has many submodules, but as a first step this fixes a
bug case in ka-clone.

Test Plan: integration tests updated, run `make tests`

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: sean

Differential Revision: https://phabricator.khanacademy.org/D22107