Don't pull `pre-push.sample` into the pre-push hook
Summary:
Some folks appear to have a `pre-push.sample` file in their
`webapp/.git` directory that prevents pushing commits that start with
`WIP`. I imagine this was installed by default with some git versions
or something? It doesn't appear to be ours. Starting commits with WIP
is a pretty common thing in our code review workflow, and we probably
shouldn't prevent that.

After installing the new pre-push hook that pulls in `pre-push.*`, this
sample hook becomes active and prevents commits with `WIP`. This commit
changes it so that we ignore `pre-push.sample` in particular. I guess
another approach would be to ensure that `.git/hooks` is cleaned out
before installing hooks with ka-clone, but I'm not sure if people have
personal hooks that that would disrupt, so I went with this approach to
be minimally invasive.

Test Plan:
- in webapp, prior to installing the updated hook:
- `touch colin_push_hook_test`
- `git add colin_push_hook_test`
- `git commit -m "WIP" -n`
- `git push -u origin HEAD:colin_push_hook_test`
- verify this commit is rejected
- install the new hook from this commit with
  `ka-clone --repair --protect-master` in the webapp dir
- `git push -u origin HEAD:colin_push_hook_test`
- verify this commit is allowed
- verify I saw `khan-linter: all lint checks passed`, indicating that
  the lint hook still ran

Reviewers: mattdunnrankin, csilvers

Reviewed By: mattdunnrankin, csilvers

Differential Revision: https://phabricator.khanacademy.org/D41640