Add a version arg for use by Repo Rangers (#24)
## Summary:
[See the Slack thread here for more context](https://khanacademy.slack.com/archives/C06P903SXV3/p1746624684972699).

Before, to see if we are on the right version of `ka-clone`, Repo Rangers would check if the flags it needed were present. If they weren't, rrs would just need to update ka-clone to the latest version.

But now we are making a change to ka-clone that end up *not changing* any of ka-clone's flags (it just toggles a default), so for this change, we can no longer rely on the presence/absence of flags to determine if ka-clone is up to date. Instead, we will *also* need use a --version flag to check the version of ka-clone.

Related PRs:
- https://github.com/Khan/khan-dotfiles/pull/145
- https://github.com/Khan/ka-clone/pull/23
- https://github.com/Khan/ka-clone/pull/24 👈 you are here
- https://github.com/Khan/our-lovely-cli/pull/1000
- https://github.com/Khan/our-lovely-cli/pull/1001

Issue: FEI-6625

## Test plan:
Run `ka-clone --version` and see `1`

Author: lillialexis

Reviewers: csilvers, jeresig

Required Reviewers:

Approved By: csilvers, jeresig

Checks:

Pull Request URL: https://github.com/Khan/ka-clone/pull/24